Implement worker autoscaling
Rather than running each worker in a single separate thread, I'm now
creating a new threadpool for each worker. Since threadpools will only
spawn new threads as needed, we don't have to create our own load
pressure detection method. One downside of this is that the first
request to each worker is slower, so to deal with that (and to fix
issues with workers that fail to start) I spawn an empty task onto the
isolate's threadpool to wake up a single isolate instance.

Note that the while the threadpool will always be created, only one
instance will be created by default (i.e. autoscaling will be disabled)
until we figure out a good story for applications that rely on global
state. For now, the feature will be hidden behind a
'--experimental-autoscale' flag.